Change int to long. Use virtual class to replace switch

+++ base/base.c
@@ -2,34 +2,34 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void push(int p1, int _) {
+void push(long p1, long _) {
 	stk_push(stk_eval(p1));
 }
 
-void pop(int p1, int _) {
+void pop(long p1, long _) {
 	if (p1) {
 		*stk_at(stk_eval(p1)) = *stk_at(stk_size() - 1);
 	}
 	stk_pop();
 }
 
-void mov(int p1, int p2) {
+void mov(long p1, long p2) {
 	*stk_at(stk_eval(p1)) = stk_eval(p2);
 }
 
-void jmp(int p1, int _) {
+void jmp(long p1, long _) {
 	*stk_at(0) = stk_eval(p1);
 }
 
-void call(int p1, int _) {
-	int p = stk_eval(p1);
+void call(long p1, long _) {
+	long p = stk_eval(p1);
 	stk_push(*stk_at(0));
 	*stk_at(0) = p;
 }
 
-void ret(int p1, int _) {
-	int n = 0;
-	int p = *stk_at(stk_size() - 1);
+void ret(long p1, long _) {
+	long n = 0;
+	long p = *stk_at(stk_size() - 1);
 	if (p1) {
 		n = stk_eval(p1);
 	}
@@ -40,49 +40,49 @@ void ret(int p1, int _) {
 	}
 }
 
-static int diff;
+static long diff;
 
-void cmp(int p1, int p2) {
+void cmp(long p1, long p2) {
 	diff = stk_eval(p1) - stk_eval(p2);
 }
 
-void je(int p1, int _) {
+void je(long p1, long _) {
 	if (diff == 0) {
 		jmp(p1, 0);
 	}
 }
 
-void jne(int p1, int _) {
+void jne(long p1, long _) {
 	if (diff) {
 		jmp(p1, 0);
 	}
 }
 
-void jg(int p1, int _) {
+void jg(long p1, long _) {
 	if (diff > 0) {
 		jmp(p1, 0);
 	}
 }
 
-void jge(int p1, int _) {
+void jge(long p1, long _) {
 	if (diff >= 0) {
 		jmp(p1, 0);
 	}
 }
 
-void jl(int p1, int _) {
+void jl(long p1, long _) {
 	if (diff < 0) {
 		jmp(p1, 0);
 	}
 }
 
-void jle(int p1, int _) {
+void jle(long p1, long _) {
 	if (diff <= 0) {
 		jmp(p1, 0);
 	}
 }
 
-void Exit(int p1, int _) {
+void Exit(long p1, long _) {
 	exit(stk_eval(p1));
 }
